Create Test_Dualism.js
An (very rough) example of what sort of things one could do in a philosophical language written in a type system (A λ-calculus, a Turing Complete programming language, such as JavaScript or Python). This is the microcosmic version of what this could be; with time systems could be created similarly to video games which would allow for simulation and observation of the consequences of different philosophical ideas praxis.